requirements.txt
Flask==3.1.0
python-docx==1.1.2
fpdf==1.7.2
pdfkit==1.0.0
Jinja2==3.1.6
Werkzeug==3.1.3
MarkupSafe==3.0.2
requests==2.32.3  